fix int -> size_t

+++ losev.timofey/S1/list.h
@@ -90,7 +90,7 @@ private:
 
 struct NamedList {
   std::string name;
-  List<int> numbers;
+  List<size_t> numbers;
 };
 
 template<class T>
